Finalizado

+++ view/venda/fecharVenda.php
@@ -15,10 +15,13 @@ if ($venda !== null) {
     $produtoBll = new ProdutoBll();
     $produto = $produtoBll->findById($venda->getIdProduto());
 
-    if ($produto !== null) {
+    if ($produto !== null && $venda->getQtdeVendida() <= $produto->getQtdeEstoque()) {
         $produtoBll->updateEstoque($produto);
         $vendaBll->deleteById($venda->getId());
         header("location: lstVendas.php");
+    } else {
+        die('Produto com estoque baixo<p><a href="lstVendas.php">Voltar</a></p>');
+        
     }
 } else {
     header("location: idNotFound.php");
